requirements.txt
Django==3.2.8
python-dotenv==0.19.1
psycopg2==2.9.1